% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Data generation}
\usage{
aggregate_data(
  responses,
  variant = c("distributions", "DCPO", "Claassen", "ClaassenMulti")
)
}
\arguments{
\item{responses}{a data frame - \code{responses} element of the list returned by
\link{generate_data}}

\item{variant}{a string defining the structure of the returned data}
}
\value{
a data frame
}
\description{
Aggregates data so it fits the structure needed for the estimating routines
}
