% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coverage_scheme.R
\name{check_coverage_scheme}
\alias{check_coverage_scheme}
\title{Preparing simulation conditions}
\usage{
check_coverage_scheme(coverage, conditions)
}
\arguments{
\item{coverage}{a one-row data frame containing values of parameters that
should be hold constant across all the simulation conditions}

\item{conditions}{a data frame storing simulation conditions (other than
projects' country-years coverage) - see \link{prepare_conditions}}
}
\value{
its argument \code{coverage} (or throws an error)
}
\description{
Checks structure of a data frame storing projects' country-years coverage.
}
\details{
Data frame given by the \code{coverage} argument stores the data in a \emph{long}
format with each row representing that a given country in a given year was
covered by a given survey project. It must contain columns:
\describe{
\item{variant}{ (a string) - name of the coverage variant (must match names in the column \code{variant} of the data frame provided by the \code{conditions} argument)}
\item{year}{ (an integer) - year}
\item{project}{ (a string) - survey project name}
\item{countryGroup}{ (a string) - name of the group of country that represents a given row}
}
}
\seealso{
\link{prepare_conditions}, \link{check_conditions}
}
