% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_generated_data.R
\name{check_generated_data}
\alias{check_generated_data}
\title{Preparing simulation conditions}
\usage{
check_generated_data(conditions, coverageScheme, nIterPerCond)
}
\arguments{
\item{conditions}{a data frame with simulation conditions, typically
constructed using \link{prepare_conditions}}

\item{coverageScheme}{a data frame with country-year-survey project coverage
scheme (see \link{check_coverage_scheme})}

\item{nIterPerCond}{a positive integer - number of iterations to be run for
each condition}
}
\value{
a data frame provided by the \code{conditions} argument with additional
columns:
\describe{
\item{maxNPointDistrib}{maximum number of project-country-year-items
with one-point distributions (i.e. with only one
cell with non-zero frequency) across replications
of generating data under a given condition}
\item{q95NPointDistrib}{95. percentile as above}
\item{q90NPointDistrib}{90. percentile as above}
\item{maxPctPointDistrib}{maximum percent of project-country-year-items
as above}
\item{q95PctPointDistrib}{95. percentile as above}
\item{q90PctPointDistrib}{90. percentile as above}
\item{maxNZeroCells}{maximum number of project-country-year-items
with some zero-count cells in response distribution
across replications of generating data under a given
condition}
\item{q95NZeroCells}{95. percentile as above}
\item{q90NZeroCells}{90. percentile as above}
\item{maxPctZeroCells}{maximum percent of project-country-year-items
as above}
\item{q95PctZeroCells}{95. percentile as above}
\item{q90PctZeroCells}{90. percentile as above}
\item{maxNPointDistribClaassen}{maximum number of project-country-year-items
with one-point distributions (i.e. with
only one cell with non-zero frequency)
across replications of generating data
under a given condition}
\item{q95NPointDistribClaassen}{95. percentile as above}
\item{q90NPointDistribClaassen}{90. percentile as above}
\item{maxPctPointDistribClaassen}{maximum percent of
project-country-year-items as above}
\item{q95PctPointDistribClaassen}{95. percentile as above}
\item{q90PctPointDistribClaassen}{90. percentile as above}
}
}
\description{
Checks whether the data generated according to the given simulations contains
empty cells or cells with only several observations in distributions of
answers within a given county-year-project.
}
\examples{
check_generated_data(conditions[1:2, ], coverageScheme, 10)
}
