% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverageScheme}
\alias{coverageScheme}
\title{Country-year-survey projects coverage scheme used in our study}
\format{
A data frame with 394 observations and 4 variables.
See \link{check_coverage_scheme} for the columns description.
}
\usage{
coverageScheme
}
\description{
Country-year-survey projects coverage scheme used in our study
}
\keyword{datasets}
