% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_claassen.R
\name{format_claassen}
\alias{format_claassen}
\title{Model estimation}
\usage{
format_claassen(responses, variant = c("dichotomous", "multinomial"))
}
\arguments{
\item{responses}{a data frame with responses - typically the \code{responses}
element of a list returned by \link{generate_data}}

\item{variant}{a string indicating whether the returned data should fit
the original dichotomous Claassen's model or its multinomial extension}
}
\value{
a list
}
\description{
Transforms data on responses to the format fitting Claassen's model.
}
\details{
Adaptation of the original Claassen's code that is available at
}
