% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_country_year_parameters}
\alias{generate_country_year_parameters}
\title{Data generation}
\usage{
generate_country_year_parameters(
  nCountries,
  nYears,
  arMeanStartLB,
  arMeanStartUB,
  arMeanChangeSD,
  arVarStartLB,
  arVarStartUB,
  arVarChangeSD,
  arMeanTrendLB,
  arMeanTrendUB
)
}
\arguments{
\item{nCountries}{an integer - the number of countries
(compare \link{check_coverage_scheme})}

\item{nYears}{an integer - the number of years
(compare \link{check_coverage_scheme})}

\item{arMeanStartLB}{a numeric - see \link{prepare_conditions}}

\item{arMeanStartUB}{a numeric - see \link{prepare_conditions}}

\item{arMeanChangeSD}{a numeric - see \link{prepare_conditions}}

\item{arVarStartLB}{a numeric - see \link{prepare_conditions}}

\item{arVarStartUB}{a numeric - see \link{prepare_conditions}}

\item{arVarChangeSD}{a numeric - see \link{prepare_conditions}}

\item{arMeanTrendLB}{a numeric - see \link{prepare_conditions}}

\item{arMeanTrendUB}{a numeric - see \link{prepare_conditions}}
}
\value{
a data frame with columns \code{country}, \code{year}, \code{mean}
and \code{var}
}
\description{
Generates country-year parameters (mean and variance)
}
\seealso{
\link{generate_data}, \link{generate_latent}
}
