% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_data}
\alias{generate_data}
\title{Data generation}
\usage{
generate_data(
  pCGY,
  nRespondents,
  nCountriesPerGroup,
  variant,
  projectBiasesSD,
  nItemsProbs,
  respScaleLengthProbs,
  arMeanStartSD,
  arMeanChangeSD,
  arVarStartLB,
  arVarStartUB,
  arVarChangeSD,
  unstLoadingDefault,
  difficultyDefault,
  unstLoadingsCSD,
  unstLoadingsYSD,
  difficultyCSD,
  difficultyYSD,
  relThresholdsL,
  arMeanBounds
)
}
\arguments{
\item{pCGY}{a data frame with country-year-survey project coverage scheme
(see \link{check_coverage_scheme})}

\item{nRespondents}{an integer - the number of respondents within each
country(-year-survey project)}

\item{nCountriesPerGroup}{an integer - the number of countries to be
generated for each group of countries (compare \link{check_coverage_scheme})}

\item{variant}{a string - see \link{check_coverage_scheme}}

\item{projectBiasesSD}{a numeric - see \link{prepare_conditions}}

\item{nItemsProbs}{a string - see \link{prepare_conditions}}

\item{respScaleLengthProbs}{a string - see \link{prepare_conditions}}

\item{arMeanStartSD}{a numeric - see \link{prepare_conditions}}

\item{arMeanChangeSD}{a numeric - see \link{prepare_conditions}}

\item{arVarStartLB}{a numeric - see \link{prepare_conditions}}

\item{arVarStartUB}{a numeric - see \link{prepare_conditions}}

\item{arVarChangeSD}{a numeric - see \link{prepare_conditions}}

\item{unstLoadingDefault}{a numeric - see \link{prepare_conditions}}

\item{difficultyDefault}{a numeric - see \link{prepare_conditions}}

\item{unstLoadingsCSD}{a numeric - see \link{prepare_conditions}}

\item{unstLoadingsYSD}{a numeric - see \link{prepare_conditions}}

\item{difficultyCSD}{a numeric - see \link{prepare_conditions}}

\item{difficultyYSD}{a numeric - see \link{prepare_conditions}}

\item{relThresholdsL}{a numeric - see \link{prepare_conditions}}

\item{arMeanBounds}{a string - see \link{prepare_conditions}}
}
\value{
a list of four data frames:
\describe{
\item{countryYears}{,}
\item{projectCountryYears}{,}
\item{items}{,}
\item{responses}{.}
}
}
\description{
Generates data by calling more specific data-generating functions
}
\seealso{
\link{make_countries}, \link{generate_projects},
\link{generate_country_year_parameters}, \link{generate_latent}, \link{generate_items}
and \link{generate_responses}
}
