% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_items}
\alias{generate_items}
\title{Data generation}
\usage{
generate_items(
  pCY,
  unstLoadingDefault,
  difficultyDefault,
  unstLoadingsCSD,
  unstLoadingsYSD,
  difficultyCSD,
  difficultyYSD,
  relThresholdsL,
  unstLoadingsCL = 3 * unstLoadingsCSD,
  unstLoadingsYL = 3 * unstLoadingsYSD,
  difficultyCL = 3 * difficultyCSD,
  difficultyYL = 3 * difficultyYSD
)
}
\arguments{
\item{pCY}{a data frame with survey project-country-year parameters - see
\link{make_countries}, \link{generate_projects} and \link{generate_country_year_parameters}}

\item{unstLoadingDefault}{a numeric - see \link{prepare_conditions}}

\item{difficultyDefault}{a numeric - see \link{prepare_conditions}}

\item{unstLoadingsCSD}{a numeric - see \link{prepare_conditions}}

\item{unstLoadingsYSD}{a numeric - see \link{prepare_conditions}}

\item{difficultyCSD}{a numeric - see \link{prepare_conditions}}

\item{difficultyYSD}{a numeric - see \link{prepare_conditions}}

\item{relThresholdsL}{a numeric - see \link{prepare_conditions}}

\item{unstLoadingsCL}{a numeric - +- these will be set as limits of trunc.-norm. distribs.}

\item{unstLoadingsYL}{a numeric - +- these will be set as limits of trunc.-norm. distribs.}

\item{difficultyCL}{a numeric - +- these will be set as limits of trunc.-norm. distribs.}

\item{difficultyYL}{a numeric - +- these will be set as limits of trunc.-norm. distribs.}
}
\value{
a data frame with columns \code{project} \code{country}, \code{year}, \code{item},
\code{respScaleLength}, \code{unstLoading} and \code{thresholds}
}
\description{
Generates item parameters
}
\seealso{
\link{generate_data}, \link{generate_responses}
}
