% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_latent}
\alias{generate_latent}
\title{Data generation}
\usage{
generate_latent(pCY, nRespondents)
}
\arguments{
\item{pCY}{a data frame with survey project-country-year parameters - see
\link{make_countries}, \link{generate_projects} and \link{generate_country_year_parameters}}

\item{nRespondents}{an integer - the number of respondents within each
country(-year-survey project)}
}
\value{
a data frame with columns \code{project}, \code{country}, \code{year}, \code{respondent}
and \code{latent}
}
\description{
Generates latent trait values with project bias already incorporated
}
\seealso{
\link{generate_data}, \link{generate_responses}
}
