% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_projects}
\alias{generate_projects}
\title{Data generation}
\usage{
generate_projects(
  nProjects,
  projectBiasesSD,
  nItemsProbs,
  respScaleLengthProbs
)
}
\arguments{
\item{nProjects}{an integer - the number of survey projects
(compare \link{check_coverage_scheme})}

\item{projectBiasesSD}{a numeric - see \link{prepare_conditions}}

\item{nItemsProbs}{a named numeric vector -  compare \link{prepare_conditions}}

\item{respScaleLengthProbs}{a named numeric vector -
compare \link{prepare_conditions}}
}
\value{
a data frame with columns \code{project}, \code{bias}, \code{nItems}
and \code{respScaleLength}
}
\description{
Generates project characteristics
}
\seealso{
\link{generate_data}, \link{generate_latent}
}
