% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_responses}
\alias{generate_responses}
\title{Data generation}
\usage{
generate_responses(latent, items)
}
\arguments{
\item{latent}{a data frame returned by \link{generate_latent}}

\item{items}{a data frame returned by \link{generate_items}}
}
\value{
a data frame with columns \code{project}, \code{country}, \code{year}, \code{item},
\code{respScaleLength}, \code{respondent} and \code{response}
}
\description{
Generates responses
}
\seealso{
\link{generate_data}
}
