% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{make_countries}
\alias{make_countries}
\title{Data generation}
\usage{
make_countries(pCGY, nCountriesPerGroup)
}
\arguments{
\item{pCGY}{a data frame with country-year-survey project coverage scheme
(see \link{check_coverage_scheme})}

\item{nCountriesPerGroup}{an integer - the number of countries to be
generated for each group of countries (compare \link{check_coverage_scheme})}
}
\value{
a data frame with columns \code{project}, \code{country} and \code{year}
}
\description{
Prepares a data frame with year-project-countries
}
\seealso{
\link{generate_data}
}
