% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_stan_models.R
\name{prepare_stan_models}
\alias{prepare_stan_models}
\title{Model estimation}
\usage{
prepare_stan_models(which = c("claassen", "dcpo"))
}
\arguments{
\item{which}{optionally a character vector specifying which models should be
returned (Claassen's, Solt's or both)}
}
\value{
a list of \emph{CmdStanModel} objects
}
\description{
Returns a list of Stan models that can be used for estimation.
}
