% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_iteration.R
\name{run_iteration}
\alias{run_iteration}
\title{Simulation flow}
\usage{
run_iteration(models, coverageScheme, condition, iter, stanPars, max_redraws)
}
\arguments{
\item{models}{a list of \emph{CmdStanModel} objects - typically returned by
\link{prepare_stan_models}}

\item{coverageScheme}{a data frame with country-year-survey project coverage
scheme (see \link{check_coverage_scheme})}

\item{condition}{a one-row data frame with additional parameters specifying
data generation process that will be passed to \link{generate_data}}

\item{iter}{optionally a positive integer - number of MCMC iterations in
model estimation}

\item{stanPars}{optionally a list with additional arguments that will be
passed to the Stan models' \code{sample} method - see \link{estimate_dcpo} and
\link{estimate_claassen}}

\item{max_redraws}{Optional non-negative integer. Maximum number of retries
to re-generate data if any Claassen aggregate (Country–Year–Item–Project
proportion of affirmative answers) equals 0 or 1. If 0, no retries are made}
}
\value{
a list of three data frames:
\describe{
\item{modelSummaries}{basic model summary statistics in a \emph{long} format}
\item{countryMeans}{generated country-year means (both as returned by
\link{generate_data}, i.e. not standardized, and
standardized within the group of observed country-means
- the latter with suffix \emph{Std} added to their names)
and their estimates from the models (standardized
within the group of observed country-means)}
\item{items}{item parameters (generated, not estimated)}
}
}
\description{
Performs a single iteration of the simulation.
}
\details{
At the moment function estimates three models:
\describe{
\item{Solt's DCPO model}{some explanations and perhaps citation}
\item{Claassen's model}{some explanations and perhaps citation}
\item{Multinomial extension of the Claassen's model}{
some explanations and perhaps citation}
}
}
