% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Simulation flow}
\usage{
run_simulation(
  conditions,
  coverageScheme,
  nIterPerCond,
  suffix = "",
  iter = 1000L,
  max_redraws = 5L,
  stanPars = list()
)
}
\arguments{
\item{conditions}{a data frame with simulation conditions, typically
constructed using \link{prepare_conditions}}

\item{coverageScheme}{a data frame with country-year-survey project coverage
scheme (see \link{check_coverage_scheme})}

\item{nIterPerCond}{a positive integer - number of iterations to be run for
each condition}

\item{suffix}{optionally a string - suffix that will be added to the name
of a file storing simulation results (that will be saved to the disk)}

\item{iter}{optionally a positive integer - number of MCMC iterations in
model estimation}

\item{max_redraws}{Optional non-negative integer. Maximum number of retries
to re-generate data if any Claassen aggregate (Country–Year–Item–Project
proportion of affirmative answers) equals 0 or 1. If 0, no retries are made}

\item{stanPars}{optionally a list with additional arguments that will be
passed to the Stan models' \code{sample} method - see \link{estimate_dcpo} and
\link{estimate_claassen}}
}
\value{
(invisibly) a list of four data frames:
\describe{
\item{modelSummaries}{basic model summary statistics in a \emph{long} format}
\item{countryMeans}{generated country-year means (as returned by
\link{generate_data}, i.e. not standardized) and their
estimates from the models (standardized within the
group of observed country-means)}
\item{items}{item parameters (generated, not estimated)}
\item{itemDistributions}{distributions (counts) of responses to items for
each project-country-year-item}
}
Moreover, after completing each simulation condition-iteration function will
save to the disk file "cntYearTrends_results[suffix].RData" storing the
data frames listed above.
}
\description{
Runs the simulation.
}
\details{
See \link{run_iteration}
}
\examples{
\dontrun{
str(conditions)
str(coverageScheme)
set.seed(12345)
run_simulation(conditions[1:2, ], coverageScheme, nIterPerCond = 1L, iter = 100L, max_redraws = 10L)
}
}
