% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_claassen_results.R
\name{summarize_claassen_results}
\alias{summarize_claassen_results}
\title{Model estimation}
\usage{
summarize_claassen_results(input, output)
}
\arguments{
\item{input}{an object returned by \link{format_claassen}}

\item{output}{an object of the \emph{CmdStanFit} class}
}
\description{
Gets country mean parameters out of the fitted Claassen's model.
}
