% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dcpo_results.R
\name{summarize_dcpo_results}
\alias{summarize_dcpo_results}
\title{Model estimation}
\usage{
summarize_dcpo_results(input, output)
}
\arguments{
\item{input}{an object returned by \link[DCPOtools:format_dcpo]{DCPOtools::format_dcpo}}

\item{output}{an object of the \emph{CmdStanFit} class}
}
\description{
Gets country mean parameters out of the fitted DCPO model.
}
