% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_conditions.R
\name{prepare_conditions}
\alias{prepare_conditions}
\title{Preparing simulation conditions}
\usage{
prepare_conditions(constant, ...)
}
\arguments{
\item{constant}{a one-row data frame containing values of parameters that
should be hold constant across all the simulation conditions}

\item{...}{data frames with combinations of values of the other parameters}
}
\value{
a data frame
}
\description{
Prepares a data frame storing simulation conditions. Its arguments should be
data frames giving combinations of simulation parameters.
}
\details{
Each parameter in the list below should be given as a column in exactly one
of the data frames provided as function arguments:

\strong{Projects' characteristics}
\describe{
\item{nRespondents}{ (an integer) - number of respondents in each country}
\item{projectBiasesSD}{ (a number) - std. dev. of the centered-normal distributions from which project biases will be drawn}
\item{nItemsProbs}{ (a string with expression creating a named numeric vector summing up to 1, with names coercing to integers) - probability distribution of a project having a given number of items}
\item{respScaleLengthProbs}{ (a string with expression creating a named numeric vector summing up to 1, with names coercing to integers) - probability distribution of a project having a given length of the response scale}
}
\strong{Country-means and SDs autoregressive process parameters}
\describe{
\item{arMeanStartSD}{ (a number) - std. dev. of country means in the first year}
\item{arMeanChangeSD}{ (a number) - std. dev. of change of country means relatively to the previous year}
\item{arVarStartLB}{ (a number) - a lower bound of a uniform distrib. (of within-countries std. devs. in the first year)}
\item{arVarStartUB}{ (a number) - an upper bound of a uniform distrib. (of within-countries std. devs. in the first year)}
\item{arVarChangeSD}{ (a number) - std. dev. of a multiplicative change of within-country SD relatively to the previous year (this is parameter of a log-normal distrib.)}
}
\strong{Default item parameters}
\describe{
\item{unstLoadingDefault}{ (a number) - unstandardized loading}
\item{difficultyDefault}{ (a number) - mean of the thresholds}
\item{relThresholdsL}{ (a number) - +- this will be set as limits of the uniform distrib.}
}
\strong{Non-invariance parameters}
\describe{
\item{unstLoadingsCSD}{ (a number) - between-countries non-invariance effect std. dev.}
\item{unstLoadingsYSD}{ (a number) - temporal non-invariance effect std. dev.}
\item{difficultyCSD}{ (a number) - between-countries non-invariance effect std. dev.}
\item{difficultyYSD}{ (a number) - temporal non-invariance effect std. dev.}
}
\strong{Parameters describing the projects' coverage}
\describe{
\item{variant}{ (a string) - the name of the coverage variant}
\item{nCountriesPerGroup}{ (an integer) - the number of countries that will be created for each row of a given variant}
}
}
\examples{
constant <- data.frame(
  # projects' characteristics
  nRespondents = 1500L,
  projectBiasesSD = 0.5, # project biases SD
  nItemsProbs = "c(`1` = 0.3, `2` = 0.4, `3` = 0.3)", # number of items in a project
  respScaleLengthProbs = "c( `2` = 0.1, `4` = 0.2, `5` = 0.3, `7` = 0.3, `10` = 0.1)",
  # country-means and SDs autoregressive process parameters #####################
  arMeanStartSD = 1, # SD of country means in the first year
  arMeanChangeSD = 0.3, # SD of change of country means relatively to the previous year
  arVarStartLB = 0.6, # a lower bound of a uniform distrib.
                      # (of within-countries SDs in the first year)
  arVarStartUB = 1.4, # an upper bound of a uniform distrib.
                      # (of within-countries SDs in the first year)
  arVarChangeSD = 0, # SD of a multiplicative change of within-country SD relatively to
                     # the previous year (this is parameter of a log-normal distrib.)
  # default item parameters
  unstLoadingDefault = 0.75,
  difficultyDefault = 0, # mean of thresholds
  relThresholdsL = 0.5 # +- this will be set as limits of the uniform distrib.
)
(invariance = rbind(
  "full metric + small thresholds var." = data.frame(
    unstLoadingsCSD = 0, # between-countries non-invariance effect SD
    unstLoadingsYSD = 0, # temporal non-invariance effect SD
    difficultyCSD = 0.05, # between-countries non-invariance effect SD
    difficultyYSD = 0.2), # temporal non-invariance effect SD
  "full metric + large thresholds var." = data.frame(
    unstLoadingsCSD = 0,
    unstLoadingsYSD = 0,
    difficultyCSD = 0.2,
    difficultyYSD = 0.2),
  "small loadings var. + small thresholds var." = data.frame(
    unstLoadingsCSD = 0.05,
    unstLoadingsYSD = 0,
    difficultyCSD = 0.05,
    difficultyYSD = 0.2),
  "small loadings var. + large thresholds var." = data.frame(
    unstLoadingsCSD = 0.15,
    unstLoadingsYSD = 0,
    difficultyCSD = 0.2,
    difficultyYSD = 0.2)
))
(scheme <- rbind(
  data.frame(variant = "dense",
             nCountriesPerGroup = c(2L, 4L, 10L)), # there are 10 groups
  data.frame(variant = "sparse",
             nCountriesPerGroup = c(2L, 4L, 10L)) # there are 10 groups
))
prepare_conditions(constant, invariance, scheme)
}
\seealso{
\link{check_conditions}, \link{check_conditions_names}
}
