% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Simulation flow}
\usage{
run_simulation(conditions, coverageScheme, nIterPerCond, suffix = "")
}
\arguments{
\item{conditions}{a data frame with simulation conditions, typically
constructed using \link{prepare_conditions}}

\item{coverageScheme}{a data frame with country-year-survey project coverage
scheme (see \link{check_coverage_scheme})}

\item{nIterPerCond}{a positive integer - number of iterations to be run for
each condition}

\item{suffix}{optionally a string - suffix that will be added to the names
of files storing simulation results (that will be saved to the disk)}
}
\value{
(invisibly) a list of three data frames:
\describe{
\item{modelSummaries}{basic model summary statistics in a \emph{long} format}
\item{countryMeans}{generated country-year means (as returned by
\link{generate_data}, i.e. not standardized) and their
estimates from the models (standardized within the
group of observed country-means)}
\item{items}{item parameters (generated, not estimated)}
}
}
\description{
Runs the simulation.
}
\details{
See \link{run_iteration}
}
\examples{
\dontrun{
str(conditions)
str(coverageScheme)
set.seed(12345)
run_simulation(conditions[1:2, ], coverageScheme, nIterPerCond = 1L)
}
}
